pyproject.toml
[project]
name = "mcp-superiorapis-remote"
version = "0.1.0"
description = "MCP SuperiorAPIs Remote - Dynamic integration with Superior APIs via MCP protocol"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Superior APIs Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pytest-asyncio>=0.21.0",
]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
mcp-superiorapis-http = "mcp_superiorapis_remote.mcp_server_http:main"
mcp-superiorapis-sse = "mcp_superiorapis_remote.mcp_server_sse:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_superiorapis_remote"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
